# express-back/docs/swagger.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "哈尔滨学院校园综合服务平台 API",
    "version": "1.0.0",
    "description": "校园综合服务平台后端接口文档。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/user/forum": {
      "get": {
        "tags": [
          "User Forum"
        ],
        "summary": "公开接口",
        "operationId": "ForumController_getPosts",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true
      },
      "post": {
        "tags": [
          "User Forum"
        ],
        "summary": "需要用户认证的接口",
        "operationId": "ForumController_createPost",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/user/forum/{id}": {
      "get": {
        "tags": [
          "User Forum"
        ],
        "summary": "Get Post By Id",
        "operationId": "ForumController_getPostById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true
      },
      "put": {
        "tags": [
          "User Forum"
        ],
        "summary": "Update Post",
        "operationId": "ForumController_updatePost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User Forum"
        ],
        "summary": "Delete Post",
        "operationId": "ForumController_deletePost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/forum/{id}/comments": {
      "get": {
        "tags": [
          "User Forum"
        ],
        "summary": "Get Comments",
        "operationId": "ForumController_getComments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true
      },
      "post": {
        "tags": [
          "User Forum"
        ],
        "summary": "评论相关",
        "operationId": "ForumController_createComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}